Added features related to new subbanding in ASPFitsReader:
-nsubs: Number of contiguous output subbands in which to divide input band
-nchansub: Number of channels to include in each (contiguous) output subband
-subbw: Width of each (contiguous) output subband in MHz
-subref: Chosen reference frequency for creating subbands

This will soon (but not quite yet) replace the -addchans option.  

For example, -nsubs 1 is the same as -addchans.

In the future, will  need to add option to choose output subband frequencies explicitly, 
e.g., -subfreq 1372 1392 1396 1396 1400 1402

and so on, so that subband widths need not be equal.  Once this is done, then we can fully make -addchans obsolete.

See (upcoming) documentation for usage of these options 



git-svn-id: svn+ssh://nanograv.astro.ubc.ca/data1/svn/repos/ASPFitsReader@12 cca996d7-57bb-40df-b7ff-243d9995c488