New routine (and associated command line interpreter file) to create a template profile from scratch, or using an existing profile to which one can add new profiles.
Also has the feature of performing a noise cut in the Fourier domain, as described in Paul Demorest's thesis.

Also added is a function which check whether all elements in an array are equal to zero.  Uses by this and other routines to determine if a scan has been flagged for omission.

-- Robert Ferdman, 17 April, 2009



git-svn-id: svn+ssh://nanograv.astro.ubc.ca/data1/svn/repos/ASPFitsReader@44 cca996d7-57bb-40df-b7ff-243d9995c488